Sprint 12: Добавить Mpa и Genre, создать сервисы, контроллеры, репозитории для этих сущностей. Оптимизировать код в UserService, FilmService

+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -8,12 +8,12 @@ import java.util.Optional;
 public interface MpaStorage {
     Mpa create(Mpa mpa);
 
-    Optional<Mpa> findById(Long id);
+    Optional<Mpa> findById(Integer id);
 
     Collection<Mpa> findAll();
 
     Optional<Mpa> update(Mpa newMpa);
 
-    Optional<Mpa> removeById(Long id);
+    Optional<Mpa> removeById(Integer id);
 
 }
